GameState

Description

Module containing data structures to represent the game state, actions to modify state, and rendering functions to display state.

Documentation

data Player

Constructors

X
Player X

O
Player O

type Turn = Player

newtype Board

Constructors

Board String

data GameState board turn

Constructors

GameState board turn
GameState accepts polymorphic types so that the board and/or turn can also be represented as alternative data structures (such as a string, or a list of X for board)

dim :: Int

The length of the side of the board - default value is 3.

size :: Int

The number of cells in the board - \( \dim^2 \).

red :: Paint

The color of X.

blue :: Paint

The color of O.

initBoard :: GameState Board Turn

The state of the game upon initialization.

colorize :: Board -> String

Color X red and O blue in the terminal.

render :: Board -> String

Display the board as a user-friendly string.

>>> render (Board "XOX OO X ")
 0 | 1 | 2
-----------
 3 | 4 | 5
-----------
 6 | 7 | 8

switchTurn :: Turn -> Turn

Alternate turns between X and O

>>> switchTurn X
O

move :: GameState Board Turn -> Int -> GameState Board Turn

move marks the desired position with the symbol corresponding to the current player.

possibleMoves :: GameState Board Turn -> [Int]

possibleMoves returns a list of the remaining valid moves.

>>> possibleMoves (GameState (Board "XXXOOOXX ") O)
[8]

isStart :: GameState Board Turn -> Bool

Determine whether the game board is empty.

isWinFor :: GameState Board Turn -> Player -> Bool

Determine whether Player has achieved the victory condition.

isEnd :: GameState Board Turn -> Bool

Determine whether the game has ended.